A shit ton of stuff

+++ DVN Controller/SideMenuCtrl.h
@@ -8,7 +8,9 @@ private:
 	wxPanel* mainPanel;
 
 	wxMenu* contextMenu;
-	
+
+	DVNFileData* source;
+
 	void OnContextMenu(wxMouseEvent& e) {
 		PopupMenu(contextMenu, e.GetPosition());
 
@@ -17,7 +19,7 @@ private:
 		NameSetter* nameSetter = new NameSetter(base, "Enter scenario name", DVNFileData::ValidateName);
 		nameSetter->ShowModal();
 		if (nameSetter->ok) {
-			data->Rename(nameSetter->name);
+			source->Rename(nameSetter->name);
 			SetLabel(nameSetter->name);
 		}
 	}
@@ -27,27 +29,37 @@ private:
 		mainPanel->GetEventHandler()->ProcessEvent(eOut);
 	}
 public:
-	DVNFileData* data;
-	SideMenuCtrl(wxWindow* parent, wxPanel* mainPanel, DVNFileData* data) : wxButton(parent, wxID_ANY, data->GetName(), wxDefaultPosition, wxSize(-1, 43)) {
-		this->data = data;
+	SideMenuCtrl(wxWindow* parent, wxPanel* mainPanel, DVNFileData* source, bool subMenu) : wxButton(parent, wxID_ANY, source->GetName()) {
+		SetMinSize(FromDIP(wxSize(-1, 40)));
+
+		this->source = source;
 		this->mainPanel = mainPanel;
 		SetBackgroundColour(wxColour(255, 255, 255));
 
 		contextMenu = new wxMenu();
 
 		wxMenuItem* rename = new wxMenuItem(contextMenu, wxID_ANY, "Rename");
-		wxMenuItem* deleteItem = new wxMenuItem(contextMenu, wxID_DELETE, "Delete");
-
 		rename->SetBitmap(wxArtProvider::GetBitmap(wxART_EDIT));
-		deleteItem->SetBitmap(wxArtProvider::GetBitmap(wxART_CLOSE));
-
 		contextMenu->Append(rename);
-		contextMenu->Append(deleteItem);
-
 		contextMenu->Bind(wxEVT_MENU, &SideMenuCtrl::OnRename, this, rename->GetId());
-		contextMenu->Bind(wxEVT_MENU, &SideMenuCtrl::OnDelete, this, deleteItem->GetId());
+
+		if (subMenu) {
+			
+		}
+		else {
+			wxMenuItem* deleteItem = new wxMenuItem(contextMenu, wxID_DELETE, "Delete");
+			deleteItem->SetBitmap(wxArtProvider::GetBitmap(wxART_CLOSE));
+			contextMenu->Append(deleteItem);
+			contextMenu->Bind(wxEVT_MENU, &SideMenuCtrl::OnDelete, this, deleteItem->GetId());
+		}
 
 
 		Bind(wxEVT_RIGHT_UP, &SideMenuCtrl::OnContextMenu, this);
+	}
+
+	DVNFileData* GetSource() { return source; }
+	void ChangeSource(DVNFileData* source) {
+		this->source = source;
+		SetLabel(source->GetName());
 	}
 };
